+++ src/core/box.ts
@@ -7,7 +7,7 @@ import { BoxRange } from '../types/index';
 import {
   IDBData,
   BoxData,
-  BoxScheme,
+  BoxSchema,
   BoxDataTypes,
   UncheckedData,
   OptionalBoxData,
@@ -18,12 +18,12 @@ import {
 } from '../types';
 
 // Box
-export interface Box<S extends BoxScheme> extends BoxHandler<S>, BoxTask<S> {
+export interface Box<S extends BoxSchema> extends BoxHandler<S>, BoxTask<S> {
   new (initalData?: BoxData<S>): BoxData<S>;
 }
 
 // Transaction handlers of Box
-export interface BoxHandler<S extends BoxScheme> {
+export interface BoxHandler<S extends BoxSchema> {
   getName(): string;
   getVersion(): number;
   add(value: BoxData<S>, key?: IDBValidKey): Promise<IDBValidKey>;
@@ -35,30 +35,33 @@ export interface BoxHandler<S extends BoxScheme> {
     key: string | number | Date | ArrayBufferView | ArrayBuffer | IDBArrayKey | IDBKeyRange,
   ): Promise<void>;
   query(range?: BoxRange<S>): BoxCursorHandler<S>;
-  find(...filter: BoxFilterFunction<S>[]): BoxCursorHandler<S>;
+  filter(...predicate: BoxFilterFunction<S>[]): BoxCursorHandler<S>;
   clear(): Promise<void>;
   count(): Promise<number>;
 }
 
 // Box.task = BoxTask
-export interface BoxTask<S extends BoxScheme> {
+export interface BoxTask<S extends BoxSchema> {
   $add(value: BoxData<S>, key?: IDBValidKey): TransactionTask;
   $put(value: BoxData<S>, key?: IDBValidKey): TransactionTask;
   $delete(
     key: string | number | Date | ArrayBufferView | ArrayBuffer | IDBArrayKey | IDBKeyRange,
   ): TransactionTask;
-  $find(...filter: BoxFilterFunction<S>[]): BoxCursorTask<S>;
+  $query(range?: BoxRange<S>): TransactionCursorHandler<S>;
+  $filter(...predicate: BoxFilterFunction<S>[]): TransactionCursorHandler<S>;
 }
 
-// Box.find = () => BoxCursorHandler
-export interface BoxCursorHandler<S extends BoxScheme> {
+// Box.query = () => BoxCursorHandler
+// Box.filter = () => BoxCursorHandler
+export interface BoxCursorHandler<S extends BoxSchema> {
   get(order?: BoxCursorDirections, limit?: number): Promise<BoxData<S>[]>;
   update(value: OptionalBoxData<S>): Promise<void>;
   delete(): Promise<void>;
 }
 
-// Box.task.find = () => BoxCursorTask
-export interface BoxCursorTask<S extends BoxScheme> {
+// Box.$query = () => TransactionCursorHandler
+// Box.$filter = () => TransactionCursorHandler
+export interface TransactionCursorHandler<S extends BoxSchema> {
   update(value: OptionalBoxData<S>): TransactionTask;
   delete(): TransactionTask;
 }
@@ -66,15 +69,15 @@ export interface BoxCursorTask<S extends BoxScheme> {
 // Box Prototype
 export interface BoxPrototype {
   tx: BoxTransaction;
-  $(type: TransactionType, args?: TaskArguments<BoxScheme>): Promise<void | IDBData | IDBData[]>;
+  $(type: TransactionType, args?: TaskArguments<BoxSchema>): Promise<void | IDBData | IDBData[]>;
   pass(target: UncheckedData, strict?: boolean): void | never;
-  data<T extends BoxScheme>(initalData?: BoxData<T>): BoxData<T>;
+  data<T extends BoxSchema>(initalData?: BoxData<T>): BoxData<T>;
 }
 
 export interface BoxProperty {
-  store: string;
-  scheme: BoxScheme;
-  v: number;
+  __name: string;
+  __schema: BoxSchema;
+  __version: number;
 }
 
 export type BoxContext = BoxPrototype & BoxProperty;
@@ -116,27 +119,27 @@ const typeValidator = (type: BoxDataTypes, value: UncheckedData): boolean => {
 /**
  * Check object keys matching and data types
  *
- * 1. Target's key length is same with box scheme's key length
- * 2. Check target's keys in scheme
- * 3. Target's value types are correct with scheme
+ * 1. Target's key length is same with box schema's key length
+ * 2. Check target's keys in schema
+ * 3. Target's value types are correct with schema
  *
  * @param this Box
  * @param target Target data
  * @param strict Enable strict mode (disabled: check properties(like optinal) / enabled: +types)
  */
-function schemeValidator(this: BoxContext, target: UncheckedData, strict = true): void | never {
-  const schemeKeys = Object.keys(this.scheme);
+function schemaValidator(this: BoxContext, target: UncheckedData, strict = true): void | never {
+  const schemaKeys = Object.keys(this.__schema);
   const targetKeys = Object.keys(target);
 
   // Checking in strict mode
-  const samekeyLength = !strict || schemeKeys.length === targetKeys.length;
+  const samekeyLength = !strict || schemaKeys.length === targetKeys.length;
   const correctValueTypes =
     !strict ||
-    Object.entries(this.scheme).every(([k, v]) =>
+    Object.entries(this.__schema).every(([k, v]) =>
       typeValidator(typeof v === 'string' ? v : v.type, target[k]),
     );
 
-  if (!(samekeyLength && correctValueTypes && targetKeys.every((k) => schemeKeys.includes(k)))) {
+  if (!(samekeyLength && correctValueTypes && targetKeys.every((k) => schemaKeys.includes(k)))) {
     throw new BoxDBError('Data not valid');
   }
 }
@@ -147,9 +150,9 @@ function schemeValidator(this: BoxContext, target: UncheckedData, strict = true)
  * @param baseObject
  * @param targetObject
  */
-function createBoxData<T extends BoxScheme>(this: BoxContext, initalData?: BoxData<T>): BoxData<T> {
+function createBoxData<T extends BoxSchema>(this: BoxContext, initalData?: BoxData<T>): BoxData<T> {
   const boxData = {} as BoxData<T>;
-  Object.keys(this.scheme).forEach(
+  Object.keys(this.__schema).forEach(
     (k) => (boxData[k as keyof T] = (initalData && initalData[k]) ?? null),
   );
   return boxData;
@@ -163,9 +166,9 @@ function createBoxData<T extends BoxScheme>(this: BoxContext, initalData?: BoxDa
 function transactionExecuter(
   this: BoxContext,
   type: TransactionType,
-  args?: TaskArguments<BoxScheme>,
+  args?: TaskArguments<BoxSchema>,
 ) {
-  return this.tx.run(createTask(type, this.store, args));
+  return this.tx.run(createTask(type, this.__name, args));
 }
 
 /**
@@ -185,13 +188,13 @@ export default class BoxBuilder {
   private task: BoxTask<IDBData>;
 
   constructor(tx: BoxTransaction) {
-    this.proto = { tx, $: transactionExecuter, pass: schemeValidator, data: createBoxData };
+    this.proto = { tx, $: transactionExecuter, pass: schemaValidator, data: createBoxData };
     this.handler = {
       getName(this: BoxContext) {
-        return this.store;
+        return this.__name;
       },
       getVersion(this: BoxContext) {
-        return this.v;
+        return this.__version;
       },
       add(this: BoxContext, value, key) {
         this.pass(value);
@@ -215,7 +218,7 @@ export default class BoxBuilder {
           args: [key],
         });
       },
-      query(this: BoxContext, range: BoxRange<BoxScheme>) {
+      query(this: BoxContext, range: BoxRange<BoxSchema>) {
         return {
           get: (order, limit) => {
             return this.$(TransactionType.$GET, {
@@ -233,21 +236,21 @@ export default class BoxBuilder {
           },
         };
       },
-      find(this: BoxContext, ...filter) {
+      filter(this: BoxContext, ...predicate) {
         return {
           get: (order, limit) => {
             return this.$(TransactionType.$GET, {
               direction: order,
-              filter,
+              filter: predicate,
               limit,
             });
           },
           update: (value) => {
             this.pass(value, false);
-            return this.$(TransactionType.$UPDATE, { filter, updateValue: value });
+            return this.$(TransactionType.$UPDATE, { filter: predicate, updateValue: value });
           },
           delete: () => {
-            return this.$(TransactionType.$DELETE, { filter });
+            return this.$(TransactionType.$DELETE, { filter: predicate });
           },
         };
       },
@@ -262,23 +265,37 @@ export default class BoxBuilder {
     this.task = {
       $add(this: BoxContext, value, key) {
         this.pass(value);
-        return createTask(TransactionType.ADD, this.store, { args: [value, key] });
+        return createTask(TransactionType.ADD, this.__name, { args: [value, key] });
       },
       $put(this: BoxContext, value, key) {
         this.pass(value, false);
-        return createTask(TransactionType.PUT, this.store, { args: [value, key] });
+        return createTask(TransactionType.PUT, this.__name, { args: [value, key] });
       },
       $delete(this: BoxContext, key) {
-        return createTask(TransactionType.DELETE, this.store, { args: [key] });
+        return createTask(TransactionType.DELETE, this.__name, { args: [key] });
       },
-      $find(this: BoxContext, ...filter) {
+      $query(this: BoxContext, range) {
         return {
           update: (value) => {
             this.pass(value, false);
-            return createTask(TransactionType.$UPDATE, this.store, { filter, updateValue: value });
+            return createTask(TransactionType.$UPDATE, this.__name, { range, updateValue: value });
           },
           delete: () => {
-            return createTask(TransactionType.$DELETE, this.store, { filter });
+            return createTask(TransactionType.$DELETE, this.__name, { range });
+          },
+        };
+      },
+      $filter(this: BoxContext, ...predicate) {
+        return {
+          update: (value) => {
+            this.pass(value, false);
+            return createTask(TransactionType.$UPDATE, this.__name, {
+              filter: predicate,
+              updateValue: value,
+            });
+          },
+          delete: () => {
+            return createTask(TransactionType.$DELETE, this.__name, { filter: predicate });
           },
         };
       },
@@ -289,21 +306,21 @@ export default class BoxBuilder {
    * Create new box
    *
    * @param storeName Object store name
-   * @param scheme Data scheme
+   * @param schema Data schema
    */
-  build<S extends BoxScheme>(targetVersion: number, storeName: string, scheme: S): Box<S> {
-    const Box = function Box<S extends BoxScheme>(this: BoxContext, initalData?: BoxData<S>) {
-      // Check scheme if initial data provided
+  build<S extends BoxSchema>(targetVersion: number, storeName: string, schema: S): Box<S> {
+    const Box = function Box<S extends BoxSchema>(this: BoxContext, initalData?: BoxData<S>) {
+      // Check schema if initial data provided
       initalData && this.pass(initalData);
 
-      // Create empty(null) object or initalData based on scheme
+      // Create empty(null) object or initalData based on schema
       return this.data(initalData);
     } as unknown as Box<S>;
 
     const context = Object.create(this.proto) as BoxContext;
-    context.store = storeName;
-    context.scheme = scheme;
-    context.v = targetVersion;
+    context.__name = storeName;
+    context.__schema = schema;
+    context.__version = targetVersion;
 
     // Handlers
     const handler = Object.assign(context, this.handler, this.task);